refactor: code cleanup — merge modules, consolidate constants, scaffold intelligence
1. Merged freshness-stats.ts into freshness.ts (one module for all freshness logic)
2. Consolidated constants into src/utils/constants.ts (KNOWN_TECH_TERMS, META_FILES,
   EXCLUDED_DIRS, VOLATILE_TAGS, STABLE_TAGS)
3. Created src/intelligence/ scaffold (index.ts, types.ts) for upcoming TF-IDF tagging
4. Unexported internal-only functions (ensureBrainDir, removeObsidianLinks)
5. Updated all imports across 14 files

Build clean, all tests pass.
