+++ src/intelligence/index.ts
@@ -1,0 +1,1 @@
+// Intelligent tagging module — TF-IDF, bigrams, corpus stats
